Add Linux ppc64le support to CUDA stubs.
Update implib.so to include a commit with ppc64le support.

May fix https://github.com/google/jax/issues/19992, although I don't have such a machine so it's untested.

This is intended to be a cherry-pick of 574a844a6172b73b03ad2d2cf89d73389fea4522, so credit
to the original author.  Manually applied those changes to this version of the tree
that is prior to the reorg.  This version branches from the same commit that jax 0.4.25
references.